tests: interrupt tests are moved to separate folder
