+++ demo/ui/chainlit_ui/tests/__init__.py
@@ -1,0 +1,1 @@
+"""Chainlit demo tests."""
