+++ demo/ui/openwebui/tests/__init__.py
@@ -1,0 +1,1 @@
+"""Open WebUI demo tests."""
