+++ demo/ui/openwebui/tests/conftest.py
@@ -1,7 +1,24 @@
+from unittest.mock import AsyncMock
+
 import pytest
+
+from lgos_openwebui.functions import generic
+from lgos_openwebui.functions.generic import Pipe
+
+from .openwebui_support import model
 
 
 @pytest.fixture
 def anyio_backend() -> str:
     """Run the Open WebUI test suite on its supported async backend."""
     return "asyncio"
+
+
+@pytest.fixture
+def configured_pipe(monkeypatch: pytest.MonkeyPatch) -> Pipe:
+    monkeypatch.setattr(
+        generic,
+        "_retrieve_model",
+        AsyncMock(return_value=model()),
+    )
+    return Pipe()
