+++ tests/api/interrupt/__init__.py
@@ -1,0 +1,1 @@
+"""API interrupt test package."""
